Contents

Contents

Barzilai Borwein Method

Gradient Descent with Fixed Step Size

Lipschitz Approximation (Malitsky & Mishchenko)

Weighted Norm Damping Gradient Method (WNGrad)

Nesterov's Accelerated Gradient Descent

Gradient Descent with Diminishing Step Size

Index

Barzilai Borwein Method

OptimizationMethods.barzilai_borwein_gd — Function

barzilai_borwein_gd(optData::BarzilaiBorweinGD{T}, progData::P 
    where P <: AbstractNLPModel{T, S}) where {T,S}

Implements gradient descent with Barzilai-Borwein step size and applies the method to the optimization problem specified by progData.

Reference(s)

Barzilai and Borwein. "Two-Point Step Size Gradient Methods". IMA Journal of Numerical Analysis.

Method

Given iterates $\lbrace x_0,\ldots,x_k\rbrace$, the iterate $x_{k+1}$ is equal to $x_k - \alpha_k \nabla f(x_k)$, where $\alpha_k$ is one of two versions.

Long Step Size Version (if optData.long_stepsize==true)

If $k=0$, then $\alpha_0$ is set to optData.init_stepsize. For $k>0$,

\[\alpha_k = \frac{ \Vert x_k - x_{k-1} \Vert_2^2}{(x_k - x_{k-1})^\intercal (\nabla f(x_k) - \nabla f(x_{k-1}))}.\]

Short Step Size Version (if optData.long_stepsize==false)

If $k=0$, then $\alpha_0$ is set to optData.init_stepsize. For $k>0$,

\[\alpha_k = \frac{(x_k - x_{k-1})^\intercal (\nabla f(x_k) - \nabla f(x_{k-1}))}{\Vert \nabla f(x_k) - \nabla f(x_{k-1})\Vert_2^2}.\]

Arguments

optData::BarzilaiBorweinGD{T}, the specification for the optimization method.

progData<:AbstractNLPModel{T,S}, the specification for the optimization problem.

Warning

progData must have an initialize function that returns subtypes of AbstractPrecompute and AbstractProblemAllocate, where the latter has a grad argument.

OptimizationMethods.BarzilaiBorweinGD — Type

BarzilaiBorweinGD{T} <: AbstractOptimizerData{T}

A structure for storing data about gradient descent using the Barzilai-Borwein step size, and the progress of its application on an optimization problem.

Fields

name:String, name of the solver for reference.

init_stepsize::T, initial step size to start the method.

long_stepsize::Bool, flag for step size; if true, use the long version of Barzilai-Borwein. If false, use the short version.

threshold::T, gradient threshold. If the norm gradient is below this, then iteration stops.

max_iterations::Int64, max number of iterations (gradient steps) taken by the solver.

iter_diff::Vector{T}, a buffer for storing differences between subsequent iterate values that are used for computing the step size

grad_diff::Vector{T}, a buffer for storing differences between gradient values at adjacent iterates, which is used to compute the step size

iter_hist::Vector{Vector{T}}, a history of the iterates. The first entry corresponds to the initial iterate (i.e., at iteration 0). The k+1 entry corresponds to the iterate at iteration k.

grad_val_hist::Vector{T}, a vector for storing max_iterations+1 gradient norm values. The first entry corresponds to iteration 0. The k+1 entry correpsonds to the gradient norm at iteration k.

stop_iteration::Int64, the iteration number that the solver stopped on. The terminal iterate is saved at iter_hist[stop_iteration+1].

Constructors

BarzilaiBorweinGD(::Type{T}; x0::Vector{T}, init_stepsize::T, 
    long_stepsize::Bool, threshold::T, max_iterations::Int) where {T}

Constructs the struct for the Barzilai-Borwein optimization method

Arguments

T::DataType, specific data type used for calculations.

Keyword Arguments

x0::Vector{T}, initial point to start the solver at.

init_stepsize::T, initial step size used for the first iteration.

long_stepsize::Bool, flag for step size; if true, use the long version of Barzilai-Borwein, if false, use the short version.

threshold::T, gradient threshold. If the norm gradient is below this, then iteration is terminated.

max_iterations::Int, max number of iterations (gradient steps) taken by the solver.

Gradient Descent with Fixed Step Size

OptimizationMethods.fixed_step_gd — Function

fixed_step_gd(optData::FixedStepGD{T}, progData<:AbstractNLPModel{T,S})
    where {T,S}

Implements fixed step-size gradient descent for the desired optimization problem specified by progData.

Method

The iterates are updated according to the procedure

\[x_{k+1} = x_k - \alpha ∇f(x_k),\]

where $\alpha$ is the step size, $f$ is the objective function, and $∇f$ is the gradient function of $f$.

Arguments

optData::FixedStepGD{T}, the specification for the optimization method.

progData<:AbstractNLPModel{T,S}, the specification for the optimization problem.

Warning

progData must have an initialize function that returns subtypes of AbstractPrecompute and AbstractProblemAllocate, where the latter has a grad argument.

OptimizationMethods.FixedStepGD — Type

FixedStepGD{T} <: AbstractOptimizerData{T}

A structure for storing data about fixed step-size gradient descent, and the progress of its application on an optimization problem.

Fields

name::String, name of the solver for reference

step_size::T, the step-size selection for the optimization procedure

threshold::T, the threshold on the norm of the gradient to induce stopping

max_iterations::Int, the maximum allowed iterations

iter_hist::Vector{Vector{T}}, a history of the iterates. The first entry corresponds to the initial iterate (i.e., at iteration 0). The k+1 entry corresponds to the iterate at iteration k.

grad_val_hist::Vector{T}, a vector for storing max_iterations+1 gradient norm values. The first entry corresponds to iteration 0. The k+1 entry correpsonds to the gradient norm at iteration k

stop_iteration, iteration number that the algorithm stopped at. Iterate number stop_iteration is produced.

Constructors

FixedStepGD(::Type{T}; x0::Vector{T}, step_size::T, threshold::T, 
    max_iterations::Int) where {T}

Constructs the struct for the optimizer.

Arguments

T::DataType, specific data type for the calculations

Keyword Arguments

x0::Vector{T}, the initial iterate for the optimizers

step_size::T, the step size of the optimizer

threshold::T, the threshold on the norm of the gradient to induce stopping

max_iterations::Int, the maximum number of iterations allowed

Lipschitz Approximation (Malitsky & Mishchenko)

OptimizationMethods.lipschitz_approximation_gd — Function

lipschitz_approximation_gd(optData::FixedStepGD{T}, progData::P where P 
    <: AbstractNLPModel{T, S}) where {T, S}

Implements gradient descent with adaptive step sizes formed through a lipschitz approximation for the desired optimization problem specified by progData.

Warning

This method is designed for convex optimization problems.

References(s)

Malitsky, Y. and Mishchenko, K. (2020). "Adaptive Gradient Descent without Descent." Proceedings of the 37th International Conference on Machine Learning, in Proceedings of Machine Learning Research 119:6702-6712.

Method

The iterates are updated according to the procedure,

\[x_{k+1} = x_{k} - \alpha_k \nabla f(x_{k}),\]

where $\alpha_k$ is the step size and $\nabla f$ is the gradient function of the objective function $f$.

The step size is computed depending on $k$. When $k = 0$, the step size is set to optData.init_stepsize. When $k > 0$,

\[\alpha_k = \min\left\lbrace \sqrt{1 + \theta_{k-1}}\alpha_{k-1}, \frac{\Vert x_k - x_{k-1} \Vert}{\Vert \nabla f(x_k) - \nabla f(x_{k-1})\Vert} \right\rbrace,\]

where $\theta_0 = \inf$ and $\theta_k = \alpha_k / \alpha_{k-1}$.

Arguments

optData::LipschitzApproxGD{T}, the specification for the optimization method.

progData<:AbstractNLPModel{T,S}, the specification for the optimization problem.

Warning

progData must have an initialize function that returns subtypes of AbstractPrecompute and AbstractProblemAllocate, where the latter has a grad argument.

OptimizationMethods.LipschitzApproxGD — Type

LipschitzApproxGD{T} <: AbstractOptimizerData{T}

A structure for storing data about adaptive gradient descent using a Lipschitz Approximation scheme (AdGD), and the progress of its application on an optimization problem.

Fields

name::String, name of the solver for reference

init_stepsize::T, the initial step size for the method

prev_stepsize::T, step size used at iter - 1 when iter > 1.

theta::T, element used in the computation of the step size. See the referenced paper for more information.

lipschitz_approximation::T, help the lipschitz approximation used in the computation of the step size. See the referenced paper for more information.

threshold::T, the threshold on the norm of the gradient to induce stopping

max_iterations::Int64, the maximum allowed iterations

iter_diff::Vector{T}, a buffer for storing differences between subsequent iterate values that are used for computing the step size

grad_diff::Vector{T}, a buffer for storing differences between gradient values at adjacent iterates, which is used to compute the step size

iter_hist::Vector{Vector{T}}, a history of the iterates. The first entry corresponds to the initial iterate (i.e., at iteration 0). The k+1 entry corresponds to the iterate at iteration k.

grad_val_hist::Vector{T}, a vector for storing max_iterations+1 gradient norm values. The first entry corresponds to iteration 0. The k+1 entry corresponds to the gradient norm at iteration k

stop_iteration::Int64, iteration number that the algorithm stopped at. Iterate number stop_iteration is produced.

Constructors

LipschitzApproxGD(::Type{T}; x0::Vector{T}, init_stepsize::T, threshold::T, 
    max_iterations::Int) where {T}

Constructs the struct for the optimizer.

Arguments

T::DataType, specific data type for the calculations

Keyword Arguments

x0::Vector{T}, the initial iterate for the optimizers

init_stepsize::T, the initial step size for the method

threshold::T, the threshold on the norm of the gradient to induce stopping

max_iterations::Int, the maximum number of iterations allowed

Weighted Norm Damping Gradient Method (WNGrad)

OptimizationMethods.weighted_norm_damping_gd — Function

weighted_norm_damping_gd(optData::WeightedNormDampingGD{T}, 
    progData::P where P <: AbstractNLPModel{T, S}) where {T, S}

Method that implements gradient descent with weighted norm damping step size using the specifications in optData on the problem specified by progData.

Reference

Wu, Xiaoxia et. al. "WNGrad: Learn the Learning Rate in Gradient Descent". arxiv, https://arxiv.org/abs/1803.02865

Method

Let $\theta_k$ be the $k^{th}$ iterate, and $\alpha_k$ be the $k^{th}$ step size. The optimization method generate iterates following

\[\theta_{k + 1} = \theta_{k} - \alpha_k \nabla f(\theta_k),\]

where $\nabla f$ is the gradient of the objective function $f$.

The step size depends on the iteration number $k$. For $k = 0$, the step size $\alpha_0$ is the reciprocal of optData.init_norm_damping_factor. For $k > 0$, the step size is iteratively updated as

\[\alpha_k = \left[ \frac{1}{\alpha_{k-1}} + \Vert\dot F(\theta_k)\Vert_2^2 \alpha_{k-1} \right]^{-1}.\]

Warning

When $\alpha_0 < \Vert\dot F(\theta_0)\Vert_2^{-1}$ and a globally Lipschitz smooth objective function is used, then the method is guaranteed to find an $\epsilon$-stationary point. It is recommended then that optData.init_norm_damping_factor exceed $\Vert\dot F(\theta_0)\Vert_2$.

Arguments

optData::WeightedNormDampingGD{T}, specification for the optimization algorithm.

progData::P where P <: AbstractNLPModel{T, S}, specification for the problem.

Warning

progData must have an initialize function that returns subtypes of AbstractPrecompute and AbstractProblemAllocate, where the latter has a grad argument.

OptimizationMethods.WeightedNormDampingGD — Type

WeightedNormDampingGD{T} <: AbstractOptimizerData{T}

A mutable struct that represents gradient descent using the weighted-norm damping step size. It stores the specification for the method and records values during iteration.

Fields

name::String, name of the optimizer for recording purposes

init_norm_damping_factor::T, initial damping factor. This value's reciprocal will be the initial step size.

threshold::T, norm gradient tolerance condition. Induces stopping when norm is at most threshold.

max_iterations::Int64, max number of iterates that are produced, not including the initial iterate.

iter_hist::Vector{Vector{T}}, store the iterate sequence as the algorithm progresses. The initial iterate is stored in the first position.

grad_val_hist::Vector{T}, stores the norm gradient values at each iterate. The norm of the gradient evaluated at the initial iterate is stored in the first position.

stop_iteration::Int64, the iteration number the algorithm stopped on. The iterate that induced stopping is saved at iter_hist[stop_iteration + 1].

Constructors

WeightedNormDampingGD(::Type{T}; x0::Vector{T}, init_norm_damping_factor::T, 
    threshold::T, max_iterations::Int64) where {T}

Constructs an instance of type WeightedNormDampingGD{T}.

Arguments

T::DataType, type for data and computation

Keyword Arguments

x0::Vector{T}, initial point to start the optimization routine. Saved in iter_hist[1].

init_norm_damping_factor::T, initial damping factor, which will correspond to the reciprocoal of the initial step size.

threshold::T, norm gradient tolerance condition. Induces stopping when norm at most threshold.

max_iterations::Int64, max number of iterates that are produced, not including the initial iterate.

Nesterov's Accelerated Gradient Descent

OptimizationMethods.nesterov_accelerated_gd — Function

nesterov_accelerated_gd(optData::NesterovAcceleratedGD{T}, progData::P
    where P <: AbstractNLPModel{T, S}) where {T, S}

Implements Nesterov's Accelerated Gradient Descent as specified by optData on the problem specified by progData.

Warning

This algorithm is designed for convex problems.

Reference(s)

See Algorithm 1 of Li et. al. "Convex and Non-convex Optimization Under Generalized Smoothness". arxiv, https://arxiv.org/abs/2306.01264.

See line-search based approach in Nesterov, Yurii. 1983. “A Method for Solving the Convex Programming Problem with Convergence Rate O(1/K^2).” Proceedings of the USSR Academy of Sciences 269:543–47.

Method

Let the objective function be denoted by $F(\theta)$ and $\nabla F( \theta)$ denote its gradient function. Given $\theta_0$ and a step size $\alpha$ (equal to optData.step_size), the method produces five sequences. At $k=0$,

\[\begin{cases} B_0 &= 0 \\ z_0 &= \theta_0 \\ \Delta_0 & = 1 \\ y_0 &= \theta_0; \end{cases}\]

and, for $k\in\mathbb{N}$,

\[\begin{cases} B_{k} &= B_{k-1} + \Delta_{k-1} \\ \theta_k &= y_{k-1} - \alpha \nabla F(y_{k-1}) \\ z_k &= z_{k-1} - \alpha\Delta_{k-1}\nabla F(y_{k-1}) \\ \Delta_k &= \frac{1}{2}\left( 1 + \sqrt{4 B_{k} + 1} \right) \\ y_k &= \theta_k + \frac{\Delta_{k}}{B_k + \Delta_k + \alpha^{-1}} (z_k - \theta_k). \end{cases}\]

The iterate sequence of interest is $\lbrace \theta_k \rbrace$.

Arguments

optData::NesterovAcceleratedGD{T}, the specification for the optimization method.

progData<:AbstractNLPModel{T,S}, the specification for the optimization problem.

Warning

progData must have an initialize function that returns subtypes of AbstractPrecompute and AbstractProblemAllocate, where the latter has a grad argument.

OptimizationMethods.NesterovAcceleratedGD — Type

NesterovAcceleratedGD{T} <: AbstractOptimizerData

A structure that represents Nesterov Accelerated Gradient Descent. Stores variables related to the method, and tracks quantities as the algorithm progresses.

Fields

name:String, name of the solver for reference.

step_size::T, step size used in the method.

z::Vector{T}, buffer array for auxiliary iterate sequence

y::Vector{T}, buffer array for convex combination of iterate and auxiliary sequence

B::T, auxiliary quadratic scaling term for computing acceleration weights and step size.

threshold::T, gradient threshold. If the norm gradient is below this, then iteration stops.

max_iterations::Int64, max number of iterations (gradient steps) taken by the solver.

iter_hist::Vector{Vector{T}}, a history of the iterates. The first entry corresponds to the initial iterate (i.e., at iteration 0). The k+1 entry corresponds to the iterate at iteration k.

grad_val_hist::Vector{T}, a vector for storing max_iterations+1 gradient norm values. The first entry corresponds to iteration 0. The k+1 entry corresponds to the gradient norm at iteration k.

stop_iteration::Int64, the iteration number that the solver stopped on. The terminal iterate is saved at iter_hist[stop_iteration+1].

Constructors

    NesterovAcceleratedGD(::Type{T}; x0::Vector{T}, step_size::T,
        threshold::T, max_iterations::Int64) where {T}

Arguments

T::DataType, specific data type used for calculations.

Keyword Arguments

x0::Vector{T}, initial point to start the solver at.

step_size::T, step size used in the method.

threshold::T, gradient threshold. If the norm gradient is below this, then iteration is terminated.

max_iterations::Int, max number of iterations (gradient steps) taken by the solver.

Gradient Descent with Diminishing Step Size

OptimizationMethods.DiminishingStepGD — Type

DiminishingStepGD{T} <: AbstractOptimizerData{T}

A structure for storing data about gradient descent using diminishing step sizes, and recording the progress of its application on an optimization problem.

Fields

name::String, name of the solver for reference.

step_size_function::Function, step size function. Should take the iteration number and return the step size for that iteration.

step_size_scaling::T, factor that is multipled with the amount of the step size function.

threshold::T, gradient threshold. If the norm gradient is below this, then iteration stops.

max_iterations::Int64, max number of iterations (gradient steps) taken by the solver.

iter_hist::Vector{Vector{T}}, a history of the iterates. The first entry corresponds to the initial iterate (i.e., at iteration 0). The k+1 entry corresponds to the iterate at iteration k.

grad_val_hist::Vector{T}, a vector for storing max_iterations+1 gradient norm values. The first entry corresponds to iteration 0. The k+1 entry correpsonds to the gradient norm at iteration k.

stop_iteration::Int64, the iteration number that the solver stopped on. The terminal iterate is saved at iter_hist[stop_iteration+1].

Warning

step_size_function should take in two arguments, the data type for computation T and the iteration number. For example, calling step_size_function(Float64, 1) should return the step size as a Float64 for the iteration 1.

Constructors

DiminishingStepGD(::Type{T}; x0::Vector{T}, step_size_function::Function,
    step_size_scaling::T, threshold::T, max_iterations::Int)

Constructs the struct for the diminishing step size gradient descent method.

Arguments

T::DataType, specific data type used for calculations.

Keyword Arguments

x0::Vector{T}, initial point to start the solver at.

step_size_function::Function, step size function. Should take the iteration number and return the step size for that iteration.

step_size_scaling::T, factor that is multipled with the amount of the step size function.

threshold::T, gradient threshold. If the norm gradient is below this, then iteration is terminated.

max_iterations::Int, max number of iterations (gradient steps) taken by the solver.

OptimizationMethods.diminishing_step_gd — Function

diminishing_step_gd(optData::DiminishingStepGD{T}, progData::P
    where P <: AbstractNLPModel{T, S}) where {T, S}

Implements gradient descent with diminishing step sizes and applies the method to the optimization problem specified by progData.

Reference(s)

Patel, Vivak, and Albert Berahas. “Gradient Descent in the Absence of Global Lipschitz Continuity of the Gradients.” SIAM 6 (3): 579–846. https://doi.org/10.1137/22M1527210.

Bertsekas, Dimitri. "Nonlinear Programming". 3rd Edition, Athena Scientific, Chapter 1.

Method

Given iterates $\lbrace x_0,\ldots,x_k\rbrace$, the iterate $x_{k + 1}$ is equal to $x_k - \alpha_k \nabla f(x_k)$, where $\alpha_k$ is equal to optData.step_size_scaling * optData.step_size_function(T, k).

Step Size Function

The step size function should satisfy several conditions. First, $\alpha_k > 0$ for all $k$. Second, $\lim_{k \to \infty} \alpha_k = 0.$ Finally, $\sum_{k=0}^{\infty} \alpha_k = \infty.$ See Patel and Berahas (2024) for details.

Arguments

optData::DiminishingStepGD{T}, the specification for the optimization method.

progData<:AbstractNLPModel{T,S}, the specification for the optimization problem.

Warning

progData must have an initialize function that returns subtypes of AbstractPrecompute and AbstractProblemAllocate, where the latter has a grad argument.

Below is a list of step size functions that are in the library. The step size sequence generated by these functions, $\lbrace \alpha_k \rbrace$ satisfies $\alpha_k > 0$, $\lim_{k \to\infty} \alpha_k = 0$ and $\sum_{k=0}^\infty \alpha_k = \infty$.

OptimizationMethods.inverse_k_step_size — Function

inverse_k_step_size(::Type{T}, k::Int64)

Return the inverse of k.

Method

The step size sequence generated for $k \in \mathbb{N}$ is

\[ \alpha_k = \frac{1}{k+1},\]

when using this method.

Arguments

T::Type, data type that the computations are done in.

k::Int64, index of the step size needed.

Returns

Returns a number of type T.

OptimizationMethods.inverse_log2k_step_size — Function

inverse_log2k_step_size(::Type{T}, k::Int64) where {T}

Method

The step size sequence generated when using this method is

\[ \alpha_k = \frac{1}{\lfloor \log_2(k+1) + 1 \rfloor}\]

for $k \in \mathbb{N}$.

Arguments

T::Type, data type that the computation are done in.

k::Int64, index of the step size needed.

Returns

Returns a number of type T.

OptimizationMethods.stepdown_100_step_size — Function

stepdown_100_step_size(::Type{T}, k::Int64) where {T}

Method

The step size sequence generated is

\[ \alpha_k = \frac{1}{2^i}\]

for $k \in [ (2^i-1)100, (2^{i+1}-1)100)$.

Arguments

T::Type, data type that the computation are done in.

k::Int64, index of the step size needed.

Returns

Returns a number of type T.

Index

OptimizationMethods.BarzilaiBorweinGD

OptimizationMethods.DiminishingStepGD

OptimizationMethods.FixedStepGD

OptimizationMethods.LeastSquares

OptimizationMethods.LipschitzApproxGD

OptimizationMethods.LogisticRegression

OptimizationMethods.NesterovAcceleratedGD

OptimizationMethods.PoissonRegression

OptimizationMethods.QLLogisticSin

OptimizationMethods.WeightedNormDampingGD

OptimizationMethods.barzilai_borwein_gd

OptimizationMethods.centered_exp

OptimizationMethods.centered_shifted_log

OptimizationMethods.ddlogistic

OptimizationMethods.diminishing_step_gd

OptimizationMethods.dlinear_plus_sin

OptimizationMethods.dlogistic

OptimizationMethods.fixed_step_gd

OptimizationMethods.inverse_complimentary_log_log

OptimizationMethods.inverse_k_step_size

OptimizationMethods.inverse_log2k_step_size

OptimizationMethods.inverse_probit

OptimizationMethods.linear_plus_sin

OptimizationMethods.lipschitz_approximation_gd

OptimizationMethods.logistic

OptimizationMethods.monomial_plus_constant

OptimizationMethods.nesterov_accelerated_gd

OptimizationMethods.stepdown_100_step_size

OptimizationMethods.weighted_norm_damping_gd