Contents

Contents

Regression Problems

Regression Link Functions

Regression Variance Functions

Index

Regression Problems

OptimizationMethods.LeastSquares — Type

LeastSquares{T,S} <: AbstractNLSModel{T,S}

Implements the data structure for defining a least squares problem.

Objective Function

\[\min_{x} 0.5\Vert F(x) \Vert_2^2,\]

where

\[F(x) = A * x - b.\]

A is the coefficient matrix. b is the constant vector.

Fields

meta::NLPModelMeta{T, S}, data structure for nonlinear programming models

nls_meta::NLSMeta{T, S}, data structure for nonlinear least squares models

counters::NLSCounters, counters for nonlinear least squares models

coef::Matrix{T}, coefficient matrix, A, for least squares problem

cons::Vector{T}, constant vector, b, for least squares problem

Constructors

LeastSquares(::Type{T}; nequ=1000, nvar=50) where {T}

Constructs a least squares problems with 1000 equations and 50 unknowns, where the entries of the matrix and constant vector are independent standard Gaussian variables.

Arguments

T::DataType, specific data type of the optimization parameter

Optional Keyword Arguments

nequ::Int64=1000, the number of equations in the system

nvar::Int64=50, the number of parameters in the system

LeastSquares(design::Matrix{T}, response::Vector{T}; 
    x0 = ones(T, size(design, 2))) where {T}

Constructs a least squares problem using the design as the coef matrix and the response as the cons vector.

Arguments

design::Matrix{T}, coefficient matrix for least squares.

response::Vector{T}, constant vector for least squares.

Optional Keyword Arguments

x0::Vector{T}=ones(T, size(design, 2)), default starting point for optimization algorithms.

OptimizationMethods.LogisticRegression — Type

LogisticRegression{T,S} <: AbstractNLPModel{T,S}

Implements logistic regression problem with canonical link function. If the covariate (i.e., design) matrix and response vector are not supplied, then these are simulated.

Objective Function

Let $A$ be the covariate matrix and $b$ denote the response vector. The rows of $A$ and corresponding entry of $b$ correspond to the predictors and response for the same experimental unit. Let $A_i$ be the vector that is row i of $A$, and let $b_i$ be the i entry of $b$. Note, $b_i$ is either 0 or 1.

Let $\mu(x)$ denote a vector-valued function whose ith entry is

\[\mu_i(x) = \frac{1}{1 + \exp(-A_i^\intercal x)}.\]

If $A$ has $n$ rows (i.e., $n$ is the number of observations), then the objective function is negative log-likelihood function given by

\[F(x) = -\sum_{i=1}^n b_i \log( \mu_i(x) ) + (1 - b_i) \log(1 - \mu_i(x)).\]

Fields

meta::NLPModelMeta{T,S}, data structure for nonlinear programming models

counters::Counters, counters for a nonlinear programming model

design::Matrix{T}, the design matrix, $A$, of the logistic regression problem

response::Vector{Bool}, the response vector, $b$, of the logistic regression problem

Constructors

LogisticRegression(::Type{T}; nobs::Int64 = 1000, nvar::Int64 = 50) where T

Constructs a simulated problem where the number of observations is nobs and the dimension of the parameter is nvar. The generated design matrix's first column is all 1s. The remaining columns are independent random normal entries such that each row (excluding the first entry) has unit variance. The design matrix is stored as type Matrix{T}.

LogisticRegression(design::Matrix{T}, response::Vector{Bool};
    x0::Vector{T} = ones(T, size(design, 2)) ./ sqrt(size(design, 2))
    ) where T

Constructs a LogisticRegression problem with design matrix design and response vector response. The default initial iterate, x0 is a scaling of the vector of ones. x0 is optional.

OptimizationMethods.PoissonRegression — Type

PoissonRegression{T, S} <: AbstractNLPModel{T, S}

Implements Poisson regression with the canonical link function. If the design matrix (i.e., the covariates) and responses are not supplied, they are randomly generated.

Objective Function

Let $A$ be the design matrix, and $b$ be the responses. Each row of $A$ and corresponding entry in $b$ are the predictor and observation from one unit. The entries in $b$ must be integer valued and non-negative.

Let $A_i$ be row $i$ of $A$ and $b_i$ entry $i$ of $b$. Let

\[\mu_i(x) = \exp(A_i^\intercal x).\]

Let $n$ be the number of rows of $A$ (i.e., number of observations), then the negative log-likelihood of the model is

\[\sum_{i=1}^n \mu_i(x) - b_i (A_i^\intercal x) + C(b),\]

where $C(b)$ is a constant depending on the data. We implement the objective function to be the negative log-likelihood up to the constant term $C(b)$. That is,

\[F(x) = \sum_{i=1}^n \mu_i(x) - b_i (A_i^\intercal x).\]

Warn

Because the additive term $C(b)$ is not included in the objective function, the objective function can take on negative values.

Fields

meta::NLPModelMeta{T, S}, NLPModel struct for storing meta information for the problem

counters::Counters, NLPModel Counter struct that provides evaluations tracking.

design::Matrix{T}, covariate matrix for the problem/experiment ($A$).

response::Vector{T}, observations for the problem/experiment ($b$).

Constructors

PoissonRegression(::Type{T}; nobs::Int64 = 1000, nvar::Int64 = 50) where {T}

Construct the struct for Poisson Regression when simulated data is needed. The design matrix ($A$) and response vector $b$ are randomly generated as follows. For the design matrix, the first column is all ones, and the rest are generated according to a normal distribution where each row has been scaled to have unit variance (excluding the first column). For the response vector, let $\beta$ be the "true" relationship between the covariates and response vector for the poisson regression model, then the $i$th entry of the response vector is generated from a Poisson Distribution with rate parameter $\exp(A_i^\intercal \beta)$.

PoissonRegression(design::Matrix{T}, response::Vector{T}; 
    x0::Vector{T} = zeros(T, size(design)[2])) where {T}

Constructs the struct for Poisson Regression when the design matrix and response vector are known. The initial guess, x0 is a keyword argument that is set to all zeros by default.

!!! Remark When using this constructor, the number of rows of design must be equal to the size of response. When providing x0, the number of entries must be the same as the number of columns in design.

Regression Link Functions

OptimizationMethods.logistic — Function

logistic(η::T} where {T}

Implements

\[ \mathrm{logistic}(\eta) = \frac{1}{1 + \exp(-\eta)},\]

where T is a scalar type.

Arguments

η::T, scalar. In the regression context this is the linear effect.

OptimizationMethods.inverse_complimentary_log_log — Function

inverse_complimentary_log_log(η::T) where {T}

Implements the link function

\[ \mathrm{CLogLog}^{-1}(\eta) = 1 - \exp(-\exp(\eta)).\]

Arguments

η::T, scalar. In the regression context this is the linear effect.

OptimizationMethods.inverse_probit — Function

inverse_probit(η::T) where {T}

Implements the link function

\[ \Phi(\eta) = \frac{1}{2\pi} \exp(-.5\eta^2).\]

Arguments

η::T, scalar. In the regression context this is the linear effect.

Regression Variance Functions

OptimizationMethods.monomial_plus_constant — Function

monomial_plus_constant(μ::T, p::T, c::T) where {T}

Implements the variance function

\[ V(\mu) = (\mu^{2})^p + c.\]

See Quasi-likelihood Estimation for details.

Arguments

μ::T, scalar. In the regression context, this is the estimated mean of a datapoint.

p::T, scalar. Power applied to μ^2.

c::T, scalar. Translate the effect on variance by c.

Note

For a negative c this might not be a valid variance function. For p smaller than $.5$, the variance function is not continuously differentiable at $0$.

OptimizationMethods.linear_plus_sin — Function

linear_plus_sin(μ::T) where {T}

Implements the variance function

\[ V(\mu) = 1 + \mu + \mathrm{sin}(2\pi\mu).\]

See Quasi-likelihood Estimation for details.

Arguments

μ::T, scalar. In the regression context, this is the estimated mean of a datapoint.

OptimizationMethods.centered_exp — Function

centered_exp(μ::T, p::T, c::T) where {T}

Implements the variance function

\[ V(\mu) = \exp\left( -|\mu - c|^{2p} \right).\]

See Quasi-likelihood Estimation for details.

Arguments

μ::T, scalar. In the regression context, this is the estimated mean of a datapoint.

p::T, scalar. Power applied to (μ-c)^2.

c::T, scalar. Center where noise level is highest.

Note

For p smaller than $.5$, the variance function is not continuously differentiable at $c$.

OptimizationMethods.centered_shifted_log — Function

centered_shifted_log(μ::T, p::T, c::T) where {T}

Implements the variance function

\[ V(\mu) = \log(|\mu-c|^{2p} + 1).\]

See Quasi-likelihood Estimation for details.

Arguments

μ::T, scalar. In the regression context, this is the estimated mean of a datapoint.

p::T, scalar. Power applied to μ^2.

c::T, scalar. Center where noise level is lowest.

Note

For p smaller than $.5$, the variance function is not continuously differentiable at $c$.

Index

OptimizationMethods.BarzilaiBorweinGD

OptimizationMethods.DiminishingStepGD

OptimizationMethods.FixedStepGD

OptimizationMethods.LeastSquares

OptimizationMethods.LipschitzApproxGD

OptimizationMethods.LogisticRegression

OptimizationMethods.NesterovAcceleratedGD

OptimizationMethods.PoissonRegression

OptimizationMethods.WeightedNormDampingGD

OptimizationMethods.barzilai_borwein_gd

OptimizationMethods.centered_exp

OptimizationMethods.centered_shifted_log

OptimizationMethods.diminishing_step_gd

OptimizationMethods.fixed_step_gd

OptimizationMethods.inverse_complimentary_log_log

OptimizationMethods.inverse_k_step_size

OptimizationMethods.inverse_log2k_step_size

OptimizationMethods.inverse_probit

OptimizationMethods.linear_plus_sin

OptimizationMethods.lipschitz_approximation_gd

OptimizationMethods.logistic

OptimizationMethods.monomial_plus_constant

OptimizationMethods.nesterov_accelerated_gd

OptimizationMethods.stepdown_100_step_size

OptimizationMethods.weighted_norm_damping_gd